Revert "Fix a few ambiguous overloaded calls"
This reverts commit b7c42e83cffd8928e41e85353d3590497946fd91.

+++ Firmware/mmu2_reporting.cpp
@@ -320,7 +320,7 @@ static void FullScreenMsg(const char *pgmS, uint8_t slot){
     lcd_clear();
     lcd_puts_at_P(0, 1, pgmS);
     lcd_print(' ');
-    lcd_print(uint8_t(slot + 1));
+    lcd_print(slot + 1);
 }
 
 void FullScreenMsgCut(uint8_t slot){
